Fix rebell netsync
