+++ Content.Client/Backmen/Surgery/Wounds/Systems/ClientWoundSystem.cs
@@ -1,0 +1,8 @@
+﻿using Content.Shared.Backmen.Surgery.Wounds.Systems;
+
+namespace Content.Client.Backmen.Surgery.Wounds.Systems;
+
+public sealed class ClientWoundSystem : WoundSystem
+{
+
+}
